Change deserialization of HBase responses to not retain any memory from the read buffer. This allows reusing read buffers.