@dethi 2024-12-27
Trade-off here is a lot more small allocations (for each cells) to be able to reuse the received buffers.

I think this need more than a benchmark on the isolated function, because the benchmark will show positive result but actual usage of gohbase that is processing the results as one may be negatively impacted by the increased allocations rate.