@aaronbee 2024-12-27
> Trade-off here is a lot more small allocations (for each cells) to be able to reuse the received buffers.
> 
> I think this need more than a benchmark on the isolated function, because the benchmark will show positive result but actual usage of gohbase that is processing the results as one may be negatively impacted by the increased allocations rate.

Yeah, we are doing some testing now on a real cluster and I would like to write some more tests/benchmarks. Some thoughts:
1) We are already doing 3-4 allocations per cell thanks to the protobuf Cell value that is returned. This is adding just 1 more. And we are seeing on real clusters that the Cell struct and it's other fields are allocating more (in size in addition to count) than the []byte that is being created here! I would love to remove this Cell struct in our return, but that would be a change to the API.

2) I think this is the right thing to do anyway. The existing code is leaking the entire response buffer we are reading from HBase (technically: depends on whether we are decompressing or not whether it's the entire response buffer or just the cells). This is just wrong, IMO.